added make_info config option, output now RDZ
The 'make_index' option is now 'make_info', with expanded
functionality to handle additional output columns.

The make_info option takes multiple values, roughly the number of
output columns with a fixed order: 1, 2, 4, 7

In addition, the "completeness weight" from the MANGLE mask has now
been relegated to the info file and removed from the typical output.
(This caused confusion on the mock use).